+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Add export csv for admin batch order viewset
+
 ### Fixed
 
 - Fix CSS in default and Unicamp micro-credentials after WeasyPrint update